refactor: banner의 공통적으로 사용할 수 있는 부분을 공통 컴포넌트로 분리
