+++ src/pages/Home.tsx
@@ -1,27 +1,17 @@
-import HomeBanner from "@/components/Banner/HomeBanner";
-import Header from "@/components/Header";
 import CatetoryGrid from "@/components/Category/CategoryGrid";
 import RecommandBtn from "@/components/RecommandBtn";
 import RealTimeRanking from "@/components/RealTimeRanking";
-import Footer from "@/components/Footer";
-import { css } from "@emotion/css";
+import PageLayout from "@/components/common/layouts/PageLayout";
+import HomeBanner from "@/components/common/Banner/HomeBanner";
 
 const Home = () => {
   return (
-    <div>
-      <Header />
+    <PageLayout>
       <HomeBanner />
       <CatetoryGrid />
       <RecommandBtn />
       <RealTimeRanking />
-      <div
-        className={css`
-          margin-top: 8rem;
-        `}
-      >
-        <Footer />
-      </div>
-    </div>
+    </PageLayout>
   );
 };
